clean up and move Governor
